Rework snapcast MPRIS interface to make use of snapserver's RPC API.
By making use of the RPC API, snapcast can now reflect the snapserver stream status in the MPRIS metadata. The status is set to paused when the stream is idle, and is set to playing when the stream resumes. Sending a PAUSE event will mute the stream, PLAY will unmute. STOP will still kill the snapclient instance, but a new instance is created when the server signals that playing has begun.